add test data and move tests to different packages
