+++ src/test/groovy/com/github/h2020_5gtango/vnv/lcm/controller/NetworkControllerTest.groovy
@@ -1,13 +1,16 @@
-package com.github.h2020_5gtango.vnv.lcm.scheduler
+package com.github.h2020_5gtango.vnv.lcm.controller
 
 import com.github.h2020_5gtango.vnv.lcm.restmock.TestCatalogueMock
 import com.github.h2020_5gtango.vnv.lcm.restmock.TestExecutionEngineMock
 import com.github.h2020_5gtango.vnv.lcm.restmock.TestPlatformManagerMock
 import com.github.h2020_5gtango.vnv.lcm.restmock.TestResultRepositoryMock
 import com.github.mrduguo.spring.test.AbstractSpec
+import junit.framework.TestSuite
 import org.springframework.beans.factory.annotation.Autowired
 
-class PackageControllerTest extends AbstractSpec {
+class NetworkControllerTest extends AbstractSpec {
+
+    final def NETWORK_SERVICE_ID = 'input0ns-f213-4fae-8d3f-04358e1e1445'
 
     @Autowired
     TestPlatformManagerMock testPlatformManagerMock
@@ -21,42 +24,42 @@ class PackageControllerTest extends AbstractSpec {
     @Autowired
     TestResultRepositoryMock testResultRepositoryMock
 
-    void "schedule single Package with 1 NetworkService and 2 TestSuite should produce successfully 1 Result for 1 testPlan"() {
+    void "schedule single NetworkService should produce successfully 1 Result for 1 testPlan"() {
 
         when:
-        def entity = postForEntity('/tng-vnv-lcm/api/v1/schedulers',
-                [
-                        "uuid": "input0pa-75f5-4231-90c8-12ec80a7oi97",
-                        "network_services":
-                                [
-                                        ["uuid": "input0ns-f213-4fae-8d3f-04358e1e1445"],
-                                ],
-
-                        "test_suites":
-                                [
-                                        ["uuid": "input0ts-75f5-4ca1-90c8-12ec80a79821"],
-                                        ["uuid": "9bbbd636-75f5-4ca1-90c8-12ec80a79821"],
-                                ]
-
-                ]
+        def entity = postForEntity('/tng-vnv-lcm/api/v1/schedulers/services',
+                ["service_uuid": NETWORK_SERVICE_ID]
                 , Void.class)
 
-        then:
-        testPlatformManagerMock.networkServiceInstances.size()==3
-        testPlatformManagerMock.networkServiceInstances.values().last().status=='TERMINATED'
 
-        testExecutionEngineMock.testSuiteResults.size()==5
+        then:
+        testPlatformManagerMock.networkServiceInstances.size()==1
+
+        testExecutionEngineMock.testSuiteResults.size()==1
         testExecutionEngineMock.testSuiteResults.values().last().status=='SUCCESS'
 
-        testResultRepositoryMock.testPlans.size()==3
+        testResultRepositoryMock.testPlans.size()==1
         testResultRepositoryMock.testPlans.values().last().status=='SUCCESS'
         testResultRepositoryMock.testPlans.values().last().networkServiceInstances.size()==1
+        testResultRepositoryMock.testPlans.values().each{testPlan ->
+            testPlan.testSuiteResults.size()==1
+        }
         testResultRepositoryMock.testPlans.values().last().testSuiteResults.last().status=='SUCCESS'
 
         cleanup:
         testPlatformManagerMock.reset()
         testExecutionEngineMock.reset()
         testResultRepositoryMock.reset()
+    }
+
+    void "retrieval of a single test suite's related testSuites should successfully all the tag related tests"() {
+        when:
+        List tss = getForEntity('/tng-vnv-lcm/api/v1/schedulers/services/{serviceUuid}/tests', List, NETWORK_SERVICE_ID).body
+        then:
+
+        tss.size() == 1
+        cleanup:
+        testPlatformManagerMock.reset()
 
     }
