+++ src/test/groovy/com/github/h2020_5gtango/vnv/lcm/controller/PackageControllerTest.groovy
@@ -1,17 +1,14 @@
-package com.github.h2020_5gtango.vnv.lcm.scheduler
+package com.github.h2020_5gtango.vnv.lcm.controller
 
-import com.github.h2020_5gtango.vnv.lcm.model.NetworkService
 import com.github.h2020_5gtango.vnv.lcm.restmock.TestCatalogueMock
 import com.github.h2020_5gtango.vnv.lcm.restmock.TestExecutionEngineMock
 import com.github.h2020_5gtango.vnv.lcm.restmock.TestPlatformManagerMock
 import com.github.h2020_5gtango.vnv.lcm.restmock.TestResultRepositoryMock
 import com.github.mrduguo.spring.test.AbstractSpec
-import junit.framework.TestSuite
 import org.springframework.beans.factory.annotation.Autowired
 
-class TestSuiteControllerTest extends AbstractSpec {
+class PackageControllerTest extends AbstractSpec {
 
-    final def TEST_SUITE_ID='input0ts-75f5-4ca1-90c8-12ec80a79821'
     @Autowired
     TestPlatformManagerMock testPlatformManagerMock
 
@@ -24,18 +21,31 @@ class TestSuiteControllerTest extends AbstractSpec {
     @Autowired
     TestResultRepositoryMock testResultRepositoryMock
 
-    void "schedule single TestSuite should produce successfully 1 Result for 1 testPlan"() {
+    void "schedule single Package with 1 NetworkService and 2 TestSuite should produce successfully 1 Result for 1 testPlan"() {
 
         when:
-        def entity = postForEntity('/tng-vnv-lcm/api/v1/schedulers/tests',
-                ["test_uuid": TEST_SUITE_ID]
+        def entity = postForEntity('/tng-vnv-lcm/api/v1/schedulers',
+                [
+                        "uuid": "input0pa-75f5-4231-90c8-12ec80a7oi97",
+                        "network_services":
+                                [
+                                        ["uuid": "input0ns-f213-4fae-8d3f-04358e1e1445"],
+                                ],
+
+                        "test_suites":
+                                [
+                                        ["uuid": "input0ts-75f5-4ca1-90c8-12ec80a79821"],
+                                        ["uuid": "9bbbd636-75f5-4ca1-90c8-12ec80a79821"],
+                                ]
+
+                ]
                 , Void.class)
 
         then:
         testPlatformManagerMock.networkServiceInstances.size()==3
         testPlatformManagerMock.networkServiceInstances.values().last().status=='TERMINATED'
 
-        testExecutionEngineMock.testSuiteResults.size()==3
+        testExecutionEngineMock.testSuiteResults.size()==5
         testExecutionEngineMock.testSuiteResults.values().last().status=='SUCCESS'
 
         testResultRepositoryMock.testPlans.size()==3
@@ -50,15 +60,4 @@ class TestSuiteControllerTest extends AbstractSpec {
 
     }
 
-    void "retrieval of a single network service's related NetworkServices should successfully all the tag related services "() {
-        when:
-        List nss = getForEntity('/tng-vnv-lcm/api/v1/schedulers/tests/{testUuid}/services', List,TEST_SUITE_ID).body
-        then:
-
-        nss.size() == 3
-
-        cleanup:
-        testPlatformManagerMock.reset()
-
-    }
 }
