+++ src/test/groovy/com/github/h2020_5gtango/vnv/lcm/controller/TestSuiteControllerTest.groovy
@@ -1,6 +1,6 @@
-package com.github.h2020_5gtango.vnv.lcm.scheduler
+package com.github.h2020_5gtango.vnv.lcm.controller
 
-import com.github.h2020_5gtango.vnv.lcm.restmock.DataMock
+import com.github.h2020_5gtango.vnv.lcm.model.NetworkService
 import com.github.h2020_5gtango.vnv.lcm.restmock.TestCatalogueMock
 import com.github.h2020_5gtango.vnv.lcm.restmock.TestExecutionEngineMock
 import com.github.h2020_5gtango.vnv.lcm.restmock.TestPlatformManagerMock
@@ -9,10 +9,9 @@ import com.github.mrduguo.spring.test.AbstractSpec
 import junit.framework.TestSuite
 import org.springframework.beans.factory.annotation.Autowired
 
-class NetworkControllerTest extends AbstractSpec {
+class TestSuiteControllerTest extends AbstractSpec {
 
-    final def NETWORK_SERVICE_ID = 'input0ns-f213-4fae-8d3f-04358e1e1445'
-
+    final def TEST_SUITE_ID='input0ts-75f5-4ca1-90c8-12ec80a79821'
     @Autowired
     TestPlatformManagerMock testPlatformManagerMock
 
@@ -25,43 +24,41 @@ class NetworkControllerTest extends AbstractSpec {
     @Autowired
     TestResultRepositoryMock testResultRepositoryMock
 
-    void "schedule single NetworkService should produce successfully 1 Result for 1 testPlan"() {
+    void "schedule single TestSuite should produce successfully 1 Result for 1 testPlan"() {
 
         when:
-        def entity = postForEntity('/tng-vnv-lcm/api/v1/schedulers/services',
-                ["service_uuid": NETWORK_SERVICE_ID]
+        def entity = postForEntity('/tng-vnv-lcm/api/v1/schedulers/tests',
+                ["test_uuid": TEST_SUITE_ID]
                 , Void.class)
 
-
         then:
-        testPlatformManagerMock.networkServiceInstances.size()==1
+        testPlatformManagerMock.networkServiceInstances.size()==3
+        testPlatformManagerMock.networkServiceInstances.values().last().status=='TERMINATED'
 
-        testExecutionEngineMock.testSuiteResults.size()==1
+        testExecutionEngineMock.testSuiteResults.size()==3
         testExecutionEngineMock.testSuiteResults.values().last().status=='SUCCESS'
 
-        testResultRepositoryMock.testPlans.size()==1
+        testResultRepositoryMock.testPlans.size()==3
         testResultRepositoryMock.testPlans.values().last().status=='SUCCESS'
         testResultRepositoryMock.testPlans.values().last().networkServiceInstances.size()==1
-        testResultRepositoryMock.testPlans.values().each{testPlan ->
-            testPlan.testSuiteResults.size()==1
-        }
         testResultRepositoryMock.testPlans.values().last().testSuiteResults.last().status=='SUCCESS'
 
         cleanup:
         testPlatformManagerMock.reset()
         testExecutionEngineMock.reset()
         testResultRepositoryMock.reset()
+
     }
 
-    void "retrieval of a single test suite's related testSuites should successfully all the tag related tests"() {
+    void "retrieval of a single network service's related NetworkServices should successfully all the tag related services "() {
         when:
-        List tss = getForEntity('/tng-vnv-lcm/api/v1/schedulers/services/{serviceUuid}/tests', List, NETWORK_SERVICE_ID).body
+        List nss = getForEntity('/tng-vnv-lcm/api/v1/schedulers/tests/{testUuid}/services', List,TEST_SUITE_ID).body
         then:
 
-        tss.size() == 1
+        nss.size() == 3
+
         cleanup:
         testPlatformManagerMock.reset()
 
     }
-
 }
