--- src/test/groovy/com/github/h2020_5gtango/vnv/lcm/scheduler/CatalogueEventListenerTest.groovy
@@ -1,99 +1,0 @@
-package com.github.h2020_5gtango.vnv.lcm.scheduler
-
-import com.github.h2020_5gtango.vnv.lcm.restmock.TestCatalogueMock
-import com.github.h2020_5gtango.vnv.lcm.restmock.TestExecutionEngineMock
-import com.github.h2020_5gtango.vnv.lcm.restmock.TestPlatformManagerMock
-import com.github.h2020_5gtango.vnv.lcm.restmock.TestResultRepositoryMock
-import com.github.mrduguo.spring.test.AbstractSpec
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.http.HttpStatus
-import spock.lang.Ignore
-
-class CatalogueEventListenerTest extends AbstractSpec {
-
-    public static final String MULTIPLE_TEST_PLANS_PACKAGE_ID ='multiple_scheduler:test:0.0.1'
-    public static final String BAD_REQUEST_PACKAGE_ID ='error:test:0.0.1'
-
-    @Autowired
-    TestPlatformManagerMock testPlatformManagerMock
-
-    @Autowired
-    TestExecutionEngineMock testExecutionEngineMock
-
-    @Autowired
-    TestCatalogueMock testCatalogueMock
-
-    @Autowired
-    TestResultRepositoryMock testResultRepositoryMock
-
-    void 'schedule single TestSuite and single NetworkService should produce successfully 2 Result for 2 testPlan'() {
-
-        when:
-        def entity = postForEntity('/tng-vnv-lcm/api/v1/packages/on-change',
-                [
-                        event_name: UUID.randomUUID().toString(),
-                        package_id:  MULTIPLE_TEST_PLANS_PACKAGE_ID,
-                ]
-                , Void.class)
-
-        then:
-        entity.statusCode == HttpStatus.OK
-        testPlatformManagerMock.networkServiceInstances.size()==3
-        testPlatformManagerMock.networkServiceInstances.values().last().status=='TERMINATED'
-
-        testExecutionEngineMock.testSuiteResults.size()==3
-        testExecutionEngineMock.testSuiteResults.values().last().status=='SUCCESS'
-
-        testResultRepositoryMock.testPlans.size()==3
-        testResultRepositoryMock.testPlans.values().last().status=='SUCCESS'
-        testResultRepositoryMock.testPlans.values().last().networkServiceInstances.size()==1
-        testResultRepositoryMock.testPlans.values().last().testSuiteResults.last().status=='SUCCESS'
-
-        cleanup:
-        testPlatformManagerMock.reset()
-        testExecutionEngineMock.reset()
-        testResultRepositoryMock.reset()
-    }
-
-    @Ignore
-    void 'simple access on-change endpoint with empty package_id should produce BAD_REQUEST response'() {
-
-        when:
-        def entity = postForEntity('/tng-vnv-lcm/api/v1/packages/on-change',
-                [
-                        event_name: UUID.randomUUID().toString(),
-                        package_id:  '',
-                ]
-                , Void.class)
-
-        then:
-        entity.statusCode == HttpStatus.BAD_REQUEST
-
-        cleanup:
-        testPlatformManagerMock.reset()
-        testExecutionEngineMock.reset()
-        testResultRepositoryMock.reset()
-    }
-
-    @Ignore
-    void 'simple access on-change endpoint with non existing package_id should produce successfully BAD_REQUEST response'() {
-
-        when:
-        def entity = postForEntity('/tng-vnv-lcm/api/v1/packages/on-change',
-                [
-                        event_name: UUID.randomUUID().toString(),
-                        package_id:  BAD_REQUEST_PACKAGE_ID,
-                ]
-                , Void.class)
-
-        then:
-        entity.statusCode == HttpStatus.INTERNAL_SERVER_ERROR
-
-        cleanup:
-        testPlatformManagerMock.reset()
-        testExecutionEngineMock.reset()
-        testResultRepositoryMock.reset()
-    }
-
-
-}
