+++ src/test/resources/static/json/tests.json
@@ -186,5 +186,95 @@
     "updated_at": "2018-08-05T10:00:17.184+00:00",
     "username": null,
     "uuid": "fe7ec2a8-644f-4788-9aa7-bc2ff059819e"
+  },
+  {
+    "created_at": "2018-09-11T07:23:04.942+00:00",
+    "executions": "7",
+    "last_time_executed": "2018-09-12T09:15:48.410+00:00",
+    "md5": "0cc5f876dad4fa49ebab8714356a054b",
+    "signature": null,
+    "status": "active",
+    "testd": {
+      "author": "Felipe Vicens, ATOS",
+      "description": "\"Test descriptor for http benchmarking advanced proxy\"\n",
+      "descriptor_schema": "https://raw.githubusercontent.com/sonata-nfv/tng-schema/master/test-descriptor/testdescriptor-schema.yml",
+      "name": "test-http-benchmark-advanced-proxy",
+      "test_category": [
+        {
+          "category_type": "performance"
+        }
+      ],
+      "test_configuration_parameters": [
+        {
+          "content_type": "text/plain",
+          "parameter_definition": "location of the test file",
+          "parameter_name": "config_file",
+          "parameter_value": "config.cfg"
+        },
+        {
+          "content_type": "text/plain",
+          "parameter_definition": "bash script to run the http benchmark test within docker container",
+          "parameter_name": "runner_file",
+          "parameter_value": "runner.sh"
+        }
+      ],
+      "test_execution": [
+        {
+          "tag_id": "tag_010",
+          "test_tag": "proxy-advanced"
+        }
+      ],
+      "test_type": "bash",
+      "vendor": "eu.5gtango.atos",
+      "version": "0.1"
+    },
+    "updated_at": "2018-09-11T07:23:04.942+00:00",
+    "username": null,
+    "uuid": "b68dbe19-5c02-4865-8c4b-5e43ada1b67d"
+  },
+  {
+    "created_at": "2018-09-11T07:51:37.810+00:00",
+    "executions": "14",
+    "last_time_executed": "2018-09-12T13:05:35.141+00:00",
+    "md5": "090ae81a44104fd7470e92132a804215",
+    "signature": null,
+    "status": "active",
+    "testd": {
+      "author": "Felipe Vicens, ATOS",
+      "description": "\"Test descriptor for http benchmarking advanced\"\n",
+      "descriptor_schema": "https://raw.githubusercontent.com/sonata-nfv/tng-schema/master/test-descriptor/testdescriptor-schema.yml",
+      "name": "test-http-benchmark-advanced",
+      "test_category": [
+        {
+          "category_type": "performance"
+        }
+      ],
+      "test_configuration_parameters": [
+        {
+          "content_type": "text/plain",
+          "parameter_definition": "location of the test file",
+          "parameter_name": "config_file",
+          "parameter_value": "config.cfg"
+        },
+        {
+          "content_type": "text/plain",
+          "parameter_definition": "bash script to run the http benchmark test within docker container",
+          "parameter_name": "runner_file",
+          "parameter_value": "runner.sh"
+        }
+      ],
+      "test_execution": [
+        {
+          "tag_id": "tag_001",
+          "test_tag": "http-advanced"
+        }
+      ],
+      "test_type": "bash",
+      "vendor": "eu.5gtango.atos",
+      "version": "0.1"
+    },
+    "updated_at": "2018-09-11T07:51:37.810+00:00",
+    "username": null,
+    "uuid": "aa5c779a-8cc7-47a9-9112-d2ff348898b4"
   }
 ]
